--- engine.py
@@ -1,6 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class Engine:
-    volume: float
-    pistons: int
