+++ hw_2/__init__.py
@@ -1,0 +1,13 @@
+"""
+Домашнее задание №2
+Классы и модули
+"""
+from . import base, car, engine, exceptions, plane
+
+__all__ = [
+    "base",
+    "car",
+    "engine",
+    "exceptions",
+    "plane",
+]
