+++ hw_2/engine.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+
+@dataclass
+class Engine:
+    volume: float
+    pistons: int
