+++ hw_2/pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+addopts = -v
+testpaths = tests
